+++ server/src/commands/core/ping.js
@@ -1,0 +1,12 @@
+/*
+  Description: This module is only in place to supress error notices legacy sources may get
+*/
+
+exports.run = async (core, server, socket, data) => {
+  return;
+};
+
+exports.info = {
+  name: 'ping',
+  description: 'This module is only in place to supress error notices legacy sources may get'
+};