+++ server/src/commands/mod/dumb.js
@@ -1,4 +1,4 @@
-/* 
+/*
  * Description: Make a user (spammer) dumb
  * Author: simple
  */
@@ -16,7 +16,7 @@ exports.run = async (core, server, socket, data) => {
   if (typeof data.nick !== 'string') {
     return;
   }
-  
+
   let badClient = server.findSockets({ channel: socket.channel, nick: data.nick });
 
   if (badClient.length === 0) {
@@ -38,20 +38,20 @@ exports.run = async (core, server, socket, data) => {
 
     return;
   }
-  
+
   let record = core.muzzledHashes[badClient.hash] = {
       dumb:true
   }
-  
+
   if(data.allies && Array.isArray(data.allies)){
       record.allies = data.allies;
   }
-  
+
   server.broadcast({
     cmd: 'info',
     text: `${socket.nick} muzzled ${data.nick} in ${socket.channel}, userhash: ${badClient.hash}`
   }, { uType: 'mod' });
-  
+
 }
 
 exports.requiredData = ['nick'];
@@ -59,5 +59,5 @@ exports.requiredData = ['nick'];
 exports.info = {
   name: 'dumb',
   usage: 'dumb {nick} [allies...]',
-  description: 'Cleanly disable a user messages and make him dumb'
+  description: 'Globally shadow mute a connection. Optional allies array will see muted messages.'
 };